code reorg utilizing userspace folder
