+++ keyboards/keychron/q1/ansi_encoder/keymaps/mkillewald/keymap.c
@@ -16,36 +16,12 @@
 
 #include QMK_KEYBOARD_H
 #include "keychron_common.h"
-#include "keymap_user.h"
-#include "keymap_user_config.h"
-#ifdef RGB_MATRIX_ENABLE
-#    include "rgb_matrix_user.h"
-#endif
+#include "mkillewald.h"
+#include "eeprom_user_config.h"
+#include "rgb_matrix_user.h"
+#include "layers.h"
 
 // clang-format off
-
-enum my_bootloader_state {
-    BOOTLOADER_INACTIVE,
-    BOOTLOADER_PRESSED,
-    BOOTLOADER_WAIT,
-    BOOTLOADER_DO
-} bootloader_state;
-
-enum my_keycodes {
-    KC_LIGHT_TAB_TOGGLE = QK_USER_0,
-    KC_LIGHT_ALPHAS_TOGGLE,
-    KC_FN_LAYER_TRANSPARENT_KEYS_TOGGLE,
-    KC_FN_LAYER_COLOR_TOGGLE,
-    KC_LOCK_BLANK_MAC
-};
-
-#define KC_LTTOG KC_LIGHT_TAB_TOGGLE
-#define KC_LATOG KC_LIGHT_ALPHAS_TOGGLE
-#define KC_TKTOG KC_FN_LAYER_TRANSPARENT_KEYS_TOGGLE
-#define KC_FCTOG KC_FN_LAYER_COLOR_TOGGLE
-#define KC_LBMAC KC_LOCK_BLANK_MAC
-
-static bool win_mode;
 
 const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
     [MAC_BASE] = LAYOUT_ansi_82(
@@ -93,233 +69,51 @@ const uint16_t PROGMEM encoder_map[][NUM_ENCODERS][2] = {
 #endif
 
 void matrix_init_user(void) {
-#ifdef RGB_MATRIX_ENABLE
-    rgb_matrix_init_user();
-#endif
+    matrix_init_mkillewald();
 }
 
 void keyboard_post_init_user(void) {
-    user_config_read_eeprom();
-    
-#ifdef AUTOCORRECT_OFF_AT_STARTUP
-    // toggle autocorrect off at startup
-    if (autocorrect_is_enabled()) { autocorrect_toggle(); }
-#endif
+    keyboard_post_init_mkillewald();
 }
 
 void housekeeping_task_user(void) {
-    switch(bootloader_state) {
-        case BOOTLOADER_DO:
-            // bootloader was pressed two frames ago. RGB should now be off,
-            // so we can call the bootloader.
-            reset_keyboard();
-            break;
-        case BOOTLOADER_WAIT:
-            // bootloader was pressed on previous frame, we wait this frame and
-            // set flag to do bootloader at end of next frame. For some reason, my
-            // Q2 needed this extra wait frame.
-            bootloader_state = BOOTLOADER_DO;
-            break;
-        case BOOTLOADER_PRESSED:
-            // User pressed bootloader keycode and RGB was disabled earlier in this
-            // frame. However RGB changes wont take place immediately, so we set a
-            // flag here which will be caught at end of the next frame.
-            bootloader_state = BOOTLOADER_WAIT;
-            break;
-        default:
-            break;
-        }
-
     housekeeping_task_keychron();
-}
-
-bool dip_switch_update_user(uint8_t index, bool active) {
-    win_mode = (index == 0 && active ? true : false);
-    return true;
-}
-
-layer_state_t default_layer_state_set_user(layer_state_t state) {
-    switch (get_highest_layer(state)) {
-        case MAC_BASE:
-            // load mac base settings
-            rgb_matrix_enable_noeeprom();  
-            rgb_matrix_reload_from_eeprom();
-            
-            // check disable
-            if (!user_config_get_enable_mac_base()) {
-                rgb_matrix_disable_noeeprom();
-            }
-            break;
-        case WIN_BASE:
-            // load win base settings
-            rgb_matrix_enable_noeeprom();  
-            rgb_matrix_mode_noeeprom(user_config_get_mode_win_base());
-            rgb_matrix_set_speed_noeeprom(user_config_get_spd_win_base());
-            rgb_matrix_sethsv_noeeprom(user_config_get_hsv_win_base().h,
-                                       user_config_get_hsv_win_base().s,
-                                       user_config_get_hsv_win_base().v);
-                                       
-            // check disable
-            if (!user_config_get_enable_win_base()) {
-                rgb_matrix_disable_noeeprom();
-            }
-            break;
-        default:
-            break;
-    }
-    return state;
+    housekeeping_task_mkillewald();
 }
 
 bool process_record_user(uint16_t keycode, keyrecord_t *record) {
-    if (process_record_keychron(keycode, record)) {
-        switch (keycode) {
-            case QK_BOOT:
-                // We want to turn off LEDs before calling bootloader, so here
-                // we call rgb_matrix_disable_noeeprom() and set a flag because
-                // the LEDs won't be updated until the next frame.
-                if (record->event.pressed) {
-                    rgb_matrix_disable_noeeprom();
-                    bootloader_state = BOOTLOADER_PRESSED;
-                }
-                return false;  // Skip all further processing of this key
-            case RGB_TOG:
-                if (record->event.pressed) {
-                    rgb_matrix_toggle_noeeprom();
-                    if (win_mode) {
-                        user_config_toggle_enable_win_base();
-                    } else {
-                        user_config_toggle_enable_mac_base();
-                    }
-                }
-                return false;  // Skip all further processing of this key
-            case RGB_MOD:
-                if (win_mode) {
-                    if (record->event.pressed) {
-                        rgb_matrix_step_noeeprom();
-                        user_config_set_mode_win_base(rgb_matrix_get_mode());
-                    }
-                    return false;  // Skip all further processing of this key
-                } else {
-                    return true; // Allow further processing of this key
-                }
-            case RGB_RMOD:
-                if (win_mode) {
-                    if (record->event.pressed) {
-                        rgb_matrix_step_reverse_noeeprom();
-                        user_config_set_mode_win_base(rgb_matrix_get_mode());
-                    }
-                    return false;  // Skip all further processing of this key
-                } else {
-                    return true; // Allow further processing of this key
-                }
-            case RGB_HUI:
-                if (win_mode) {
-                    if (record->event.pressed) {
-                        rgb_matrix_increase_hue_noeeprom();
-                        user_config_set_hsv_win_base(rgb_matrix_get_hsv());
-                    }
-                    return false;  // Skip all further processing of this key
-                } else {
-                    return true; // Allow further processing of this key
-                }
-            case RGB_HUD:
-                if (win_mode) {
-                    if (record->event.pressed) {
-                        rgb_matrix_decrease_hue_noeeprom();
-                        user_config_set_hsv_win_base(rgb_matrix_get_hsv());
-                    }
-                    return false;  // Skip all further processing of this key
-                } else {
-                    return true; // Allow further processing of this key
-                }
-            case RGB_SAI:
-                if (win_mode) {
-                    if (record->event.pressed) {
-                        rgb_matrix_increase_sat_noeeprom();
-                        user_config_set_hsv_win_base(rgb_matrix_get_hsv());
-                    }
-                    return false;  // Skip all further processing of this key
-                } else {
-                    return true; // Allow further processing of this key
-                }
-            case RGB_SAD:
-                if (win_mode) {
-                    if (record->event.pressed) {
-                        rgb_matrix_decrease_sat_noeeprom();
-                        user_config_set_hsv_win_base(rgb_matrix_get_hsv());
-                    }
-                    return false;  // Skip all further processing of this key
-                } else {
-                    return true; // Allow further processing of this key
-                }
-            case RGB_VAI:
-                if (win_mode) {
-                    if (record->event.pressed) {
-                        rgb_matrix_increase_val_noeeprom();
-                        user_config_set_hsv_win_base(rgb_matrix_get_hsv());
-                    }
-                    return false;  // Skip all further processing of this key
-                } else {
-                    return true; // Allow further processing of this key
-                }
-            case RGB_VAD:
-                if (win_mode) {
-                    if (record->event.pressed) {
-                        rgb_matrix_decrease_val_noeeprom();
-                        user_config_set_hsv_win_base(rgb_matrix_get_hsv());
-                    }
-                    return false;  // Skip all further processing of this key
-                } else {
-                    return true; // Allow further processing of this key
-                }
-            case RGB_SPI:
-                if (win_mode) {
-                    if (record->event.pressed) {
-                        rgb_matrix_increase_speed_noeeprom();
-                        user_config_set_spd_win_base(rgb_matrix_get_speed());
-                    }
-                    return false;  // Skip all further processing of this key
-                } else {
-                    return true; // Allow further processing of this key
-                }
-            case RGB_SPD:
-                if (win_mode) {
-                    if (record->event.pressed) {
-                        rgb_matrix_decrease_speed_noeeprom();
-                        user_config_set_spd_win_base(rgb_matrix_get_speed());
-                    }
-                    return false;  // Skip all further processing of this key
-                } else {
-                    return true; // Allow further processing of this key
-                }
-            case KC_LIGHT_TAB_TOGGLE:
-                if (record->event.pressed) {
-                    user_config_toggle_caps_lock_light_tab();
-                }
-                return false;  // Skip all further processing of this key
-            case KC_LIGHT_ALPHAS_TOGGLE:
-                if (record->event.pressed) {
-                    user_config_toggle_caps_lock_light_alphas();
-                }
-                return false;  // Skip all further processing of this key
-            case KC_FN_LAYER_TRANSPARENT_KEYS_TOGGLE:
-                if (record->event.pressed) {
-                    user_config_toggle_fn_layer_transparent_keys_off();
-                }
-                return false;  // Skip all further processing of this key
-            case KC_FN_LAYER_COLOR_TOGGLE:
-                if (record->event.pressed) {
-                    user_config_toggle_fn_layer_color_enable();
-                }
-                return false;  // Skip all further processing of this key
-            case KC_LOCK_BLANK_MAC:
-                if (record->event.pressed) {
-                    send_string(SS_LCTL(SS_LGUI("q")) SS_DELAY(225) SS_TAP(X_ESC));
-                }
-                return false;  // Skip all further processing of this key
-            default:
-                return true;  // Process all other keycodes normally
-        }
-    }
+    if (!process_record_keychron(keycode, record)) { return false; }
+    if (!process_record_mkillewald(keycode, record)) { return false; }
     return true;
+}
+
+bool rgb_matrix_indicators_advanced_user(uint8_t led_min, uint8_t led_max) {
+    uint8_t current_layer = get_highest_layer(layer_state);
+    switch (current_layer) {
+        case MAC_BASE:
+        case WIN_BASE:
+#ifdef CAPS_LOCK_INDICATOR_COLOR
+            if (host_keyboard_led_state().caps_lock) {
+                rgb_matrix_set_color_by_keycode(led_min, led_max, current_layer, is_caps_lock_indicator, CAPS_LOCK_INDICATOR_COLOR);
+            }
+#endif
+#ifdef CAPS_WORD_INDICATOR_COLOR
+            if (is_caps_word_on()) {
+                rgb_matrix_set_color_by_keycode(led_min, led_max, current_layer, is_caps_word_indicator, CAPS_WORD_INDICATOR_COLOR);
+            }
+#endif
+            break;
+        case MAC_FN:
+        case WIN_FN:
+#ifdef FN1_LAYER_COLOR
+            if (user_config_get_fn_layer_color_enable()) {
+                rgb_matrix_set_color_by_keycode(led_min, led_max, current_layer, is_not_transparent, FN1_LAYER_COLOR);
+            }
+#endif
+            if (user_config_get_fn_layer_transparent_keys_off()) {
+                rgb_matrix_set_color_by_keycode(led_min, led_max, current_layer, is_transparent, RGB_OFF);
+            }
+            break;
+    }
+    return false;
 }
